refactor to module clustering and dim reduction
